rise-and-grind-de-lon16/requirements-lambda.txt
psycopg2-binary==2.9.9
jmespath==1.0.1